+++ code/cell.modules/pkg.sys/ui.dev/src/types/index.ts
@@ -1,5 +1,3 @@
-export * from './types.Host';
-
 export * from './types';
 export * from './types.event';
 
@@ -11,3 +9,6 @@ export * from './types.Actions.handler';
 export * from './types.Actions.model.item';
 export * from './types.Actions.model';
 export * from './types.Actions.render';
+
+export * from './types.Host';
+export * from './types.Hosted';
